pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "auto-responses"
version = "0.1.0"
description = "Console-first MVP for automated job applications"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "playwright>=1.52.0",
  "pyyaml>=6.0.2",
]

[project.scripts]
auto-responses = "app.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]
